Add: Feed 컴포넘트 분리 초기 세팅
