+++ src/Pages/jiyeon/Main/Comment/Comment.js
@@ -1,10 +1,12 @@
 import React, { Component } from "react";
+
 import "./Comment.scss";
 
 class Comment extends Component {
+  
   render() {
     const {id, name, comment, isLiked, deleteComment} = this.props;
-
+    
     return (
       <div className="Comment">
         <li className="comment">
